Add design, add folder "adapters"
