+++ app/src/main/java/com/example/hungrywolfes/ui/adapters/DetailsAdapter.kt
@@ -1,6 +1,5 @@
-package com.example.hungrywolfes.ui
+package com.example.hungrywolfes.ui.adapters
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
