+++ app/src/main/java/com/example/hungrywolfes/ui/adapters/SearchedImagesAdapter.kt
@@ -1,6 +1,5 @@
-package com.example.hungrywolfes.ui
+package com.example.hungrywolfes.ui.adapters
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -11,53 +10,41 @@ import com.bumptech.glide.Glide
 import com.example.hungrywolfes.R
 import com.example.hungrywolfes.network.ListMealsImages
 
+class SearchedImagesAdapter(private val clickListener: (item: String) -> Unit) : RecyclerView.Adapter<SearchedImagesAdapter.ImageViewHolder>() {
 
-class FavoriteAdapter(
-    private val onRemove: (ListMealsImages) -> Unit,
-    private val clickListener: (item: String) -> Unit
-) : RecyclerView.Adapter<FavoriteAdapter.ImageViewHolder>() {
-
-    private var listMealsImages: MutableList<ListMealsImages> = mutableListOf()
-
+    private val listDataImages: MutableList<ListMealsImages> = mutableListOf()
     class ImageViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
-        var mealsImages: ImageView = view.findViewById(R.id.meal_image)
-        var mealName: TextView = view.findViewById(R.id.meal_name)
+        var mealsImages = view.findViewById<ImageView>(R.id.meal_image)
+        val mealDescription = view.findViewById<TextView>(R.id.meal_description)
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ImageViewHolder {
         val layout = LayoutInflater
             .from(parent.context)
-            .inflate(R.layout.item_favorites, parent, false)
+            .inflate(R.layout.item_searched_photos, parent, false)
         return ImageViewHolder(layout)
     }
 
     override fun onBindViewHolder(holder: ImageViewHolder, position: Int) {
-        val photo = listMealsImages[position]
-
+        val photo = listDataImages[position]
         Glide.with(holder.mealsImages)
             .load(photo.strMealThumb)
             .placeholder(R.drawable.loading_img)
             .error(R.drawable.ic_connection_error)
             .into(holder.mealsImages)
-        holder.mealName.text = photo.strMeal
-        holder.itemView.setOnLongClickListener {
-            listMealsImages.remove(photo)
-            onRemove(photo)
-            notifyDataSetChanged()
-            true
-        }
         holder.itemView.setOnClickListener {
-            clickListener(photo.idMeal)
+           clickListener(photo.idMeal)
         }
-    }
+        holder.mealDescription.text =  photo.strMeal
 
-    fun setDataImages(dataImages: MutableList<ListMealsImages>) {
-        this.listMealsImages.clear()
-        this.listMealsImages.addAll(dataImages)
+    }
+    fun setDataImages(dataImages: List<ListMealsImages>) {
+        this.listDataImages.clear()
+        this.listDataImages.addAll(dataImages)
         notifyDataSetChanged()
     }
 
     override fun getItemCount(): Int {
-        return listMealsImages.size
+        return listDataImages.size
     }
 }
