+++ app/src/main/java/com/example/hungrywolfes/ui/favoritesScreen/FavoriteFragment.kt
@@ -12,7 +12,7 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.RecyclerView
 import com.example.hungrywolfes.R
 import com.example.hungrywolfes.databinding.FragmentFavoriteBinding
-import com.example.hungrywolfes.ui.FavoriteAdapter
+import com.example.hungrywolfes.ui.adapters.FavoriteAdapter
 
 class FavoriteFragment : Fragment() {
     private lateinit var binding: FragmentFavoriteBinding
