+++ app/src/main/java/com/example/hungrywolfes/ui/internetScreen/InternetFragment.kt
@@ -1,0 +1,26 @@
+package com.example.hungrywolfes.ui.internetScreen
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.hungrywolfes.R
+
+
+
+class InternetFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_internet, container, false)
+    }
+
+
+}
