+++ app/src/main/java/com/example/hungrywolfes/ui/overview/HomeFragment.kt
@@ -1,7 +1,6 @@
 package com.example.hungrywolfes.ui.overview
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -13,8 +12,8 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.RecyclerView
 import com.example.hungrywolfes.R
 import com.example.hungrywolfes.databinding.HomeBinding
-import com.example.hungrywolfes.ui.CategoryAdapter
-import com.example.hungrywolfes.ui.ImagesAdapter
+import com.example.hungrywolfes.ui.adapters.CategoryAdapter
+import com.example.hungrywolfes.ui.adapters.ImagesAdapter
 
 class HomeFragment : Fragment() {
