+++ app/src/main/res/drawable/icon_internet.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="134dp"
+    android:height="122dp"
+    android:viewportWidth="134"
+    android:viewportHeight="122">
+  <path
+      android:pathData="M67,114.333C70.682,114.333 73.667,111.349 73.667,107.667C73.667,103.985 70.682,101 67,101C63.318,101 60.333,103.985 60.333,107.667C60.333,111.349 63.318,114.333 67,114.333Z"
+      android:fillColor="#C7C7C7"/>
+  <path
+      android:pathData="M69.933,54.333L57.267,41.733L40.867,25.466L30.533,15.133L18.4,2.933C17.778,2.311 17.041,1.818 16.228,1.482C15.416,1.146 14.546,0.972 13.667,0.972C12.788,0.972 11.917,1.146 11.105,1.482C10.293,1.818 9.555,2.311 8.933,2.933C7.678,4.188 6.973,5.891 6.973,7.666C6.973,9.442 7.678,11.144 8.933,12.4L18.133,21.533L27.867,31.266L42.733,46.2L53.067,56.466L71.333,74.733L89.933,93.4L115.6,119.066C116.22,119.691 116.957,120.187 117.77,120.526C118.582,120.864 119.453,121.038 120.333,121.038C121.213,121.038 122.085,120.864 122.897,120.526C123.71,120.187 124.447,119.691 125.067,119.066C125.692,118.447 126.188,117.709 126.526,116.897C126.864,116.084 127.039,115.213 127.039,114.333C127.039,113.453 126.864,112.582 126.526,111.769C126.188,110.957 125.692,110.219 125.067,109.6L69.933,54.333ZM131.8,33.866C114.436,17.028 91.188,7.628 67,7.666C56.937,7.694 46.943,9.337 37.4,12.533L48.2,23.333C61.279,20.065 74.975,20.187 87.993,23.686C101.012,27.186 112.923,33.948 122.6,43.333C123.836,44.524 125.484,45.193 127.2,45.2C128.098,45.196 128.986,45.011 129.811,44.656C130.636,44.301 131.38,43.783 132,43.133C133.175,41.864 133.811,40.188 133.774,38.459C133.737,36.73 133.029,35.083 131.8,33.866ZM12.467,25.333C8.85,27.934 5.419,30.786 2.2,33.866C1.075,35.117 0.465,36.748 0.491,38.43C0.518,40.112 1.18,41.722 2.344,42.937C3.508,44.152 5.088,44.881 6.768,44.98C8.448,45.078 10.102,44.537 11.4,43.466C14.725,40.325 18.291,37.449 22.067,34.866L12.467,25.333ZM100.333,68.6C100.951,69.23 101.687,69.731 102.5,70.074C103.312,70.417 104.185,70.596 105.067,70.6C106.37,70.595 107.642,70.208 108.728,69.488C109.813,68.767 110.664,67.745 111.175,66.546C111.685,65.348 111.834,64.026 111.601,62.744C111.369,61.462 110.766,60.276 109.867,59.333C104.306,53.571 97.649,48.981 90.287,45.833C82.925,42.684 75.007,41.041 67,41H65.933L81.6,56.666C88.714,59.078 95.14,63.172 100.333,68.6ZM36.533,49.4C31.875,52.113 27.612,55.456 23.867,59.333C22.638,60.633 21.976,62.367 22.026,64.155C22.076,65.943 22.834,67.637 24.133,68.866C25.433,70.095 27.167,70.757 28.955,70.707C30.743,70.657 32.438,69.899 33.667,68.6C37.551,64.739 42.058,61.561 47,59.2L36.533,49.4ZM43.867,83.666C42.594,84.895 41.861,86.579 41.83,88.349C41.814,89.225 41.971,90.095 42.292,90.91C42.613,91.725 43.092,92.469 43.7,93.1C44.308,93.73 45.035,94.234 45.839,94.584C46.642,94.933 47.506,95.121 48.382,95.137C50.151,95.168 51.86,94.495 53.133,93.266C56.565,90.101 60.979,88.209 65.639,87.905C70.298,87.602 74.92,88.906 78.733,91.6L61.933,74.8C55.141,75.823 48.832,78.92 43.867,83.666Z"
+      android:fillColor="#C7C7C7"/>
+</vector>
