+++ app/src/main/res/values/strings.xml
@@ -16,4 +16,9 @@
     <string name="favorites">Favorites</string>
     <string name="terms_and_conditions">Terms and conditions</string>
     <string name="my_favorites">My favorites</string>
+    
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="no_internet_connection">No internet Connection</string>
+    <string name="check_internet_message">Your internet connection is currently not available please check or try again.</string>
+    <string name="try_again">Try again</string>
 </resources>
